add playstate as main class for game states major refactoring for survival and classic states remove shadow upgrade bonus makes better energy cost and reload time remove bullet waving increased asteroid radius change HUD add hub state for survival remove not used assets change bullet color

+++ core/src/com/swar/game/Models/Ship.java
@@ -10,7 +10,7 @@ public class Ship implements Killable, Moveable{
 
     private float speed;
 
-    private int energy = 200;
+    private float energy = 200;
     public int armor;
     private String shipSprite;
     private String Sprite;
@@ -131,11 +131,11 @@ public class Ship implements Killable, Moveable{
         this.shipSprite = shipSprite;
     }
 
-    public int getEnergy() {
+    public float getEnergy() {
         return energy;
     }
 
-    public void setEnergy(int energy) {
+    public void setEnergy(float energy) {
         this.energy = energy;
     }
